Create move settings form to <SettingsForm> component
